+++ requirements.txt
@@ -4,3 +4,4 @@ accelerate
 torchtext
 wandb
 pandas
+scikit-learn
